tmp/requirements.txt


# pip install opencv-python
# pip install pytesseract
# pip install tqdm
# pip install python-dotenv


pytesseract~=0.3.10
opencv-python~=4.5.5.62
tqdm~=4.64.1

python-dotenv~=1.0.0


#
# matplotlib~=3.7.1
# aiogram~=2.25.1
#
# numpy~=1.24.2
# SciencePlots~=2.1.0